+++ R/retry.R
@@ -18,6 +18,9 @@
 #'   full jitter - this means that each request will randomly wait between 0
 #'   and \code{pause_base * 2 ^ attempt} seconds, up to a maximum of
 #'   \code{pause_cap} seconds.
+#' @param pause_min Minimum time to with in the backoff; generally
+#'   only necessary if you need pauses less than one second (which may
+#'   not be kind to the server, use with caution!).
 #' @param quiet If \code{FALSE}, will print a message displaying how long
 #'   until the next request.
 #' @param terminate_on Optional vector of numeric HTTP status codes that if found
@@ -39,7 +42,7 @@
 #' }
 RETRY <- function(verb, url = NULL, config = list(), ...,
                   body = NULL, encode = c("multipart", "form", "json", "raw"),
-                  times = 3, pause_base = 1, pause_cap = 60,
+                  times = 3, pause_base = 1, pause_cap = 60, pause_min = 1,
                   handle = NULL, quiet = FALSE, terminate_on = NULL) {
   stopifnot(is.numeric(times), length(times) == 1L)
   stopifnot(is.numeric(pause_base), length(pause_base) == 1L)
@@ -52,7 +55,8 @@ RETRY <- function(verb, url = NULL, config = list(), ...,
 
   i <- 1
   while (!retry_should_terminate(i, times, resp, terminate_on)) {
-    backoff_full_jitter(i, resp, pause_base, pause_cap, quiet = quiet)
+    backoff_full_jitter(i, resp, pause_base, pause_cap, pause_min,
+                        quiet = quiet)
 
     i <- i + 1
     resp <- tryCatch(request_perform(req, hu$handle$handle), error = function(e) e)
@@ -77,8 +81,9 @@ retry_should_terminate <- function(i, times, resp, terminate_on) {
   }
 }
 
-backoff_full_jitter <- function(i, resp, pause_base = 1, pause_cap = 60, quiet = FALSE) {
-  length <- ceiling(stats::runif(1, max = min(pause_cap, pause_base * (2 ^ i))))
+backoff_full_jitter <- function(i, resp, pause_base = 1, pause_cap = 60,
+                                pause_min = 1, quiet = FALSE) {
+  length <- max(pause_min, stats::runif(1, max = min(pause_cap, pause_base * (2 ^ i))))
   if (!quiet) {
     if (inherits(resp, "error")) {
       error_description <- gsub("[\n\r]*$", "\n", as.character(resp))
@@ -87,7 +92,7 @@ backoff_full_jitter <- function(i, resp, pause_base = 1, pause_cap = 60, quiet =
       error_description <- ""
       status <- status_code(resp)
     }
-    message(error_description, "Request failed [", status, "]. Retrying in ", length, " seconds...")
+    message(error_description, "Request failed [", status, "]. Retrying in ", round(length, 1L), " seconds...")
   }
   Sys.sleep(length)
 }